poorpigssolver.py
This Python program calculates the minimum number of pigs required to identify a poisoned bucket among a given set of buckets. The problem is defined by parameters such as the number of buckets, the time it takes for a pig to die (minutesToDie), and the time for a test (minutesToTest). The code employs a mathematical approach to find the minimum number of pigs needed to optimize the testing process, ensuring the most efficient use of resources in detecting the poisoned bucket. Users can input their own parameters for customized scenarios.